requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
et-xmlfile==1.1.0
flake8==5.0.4
flake8-docstrings==1.6.0
frozenlist==1.3.1
geographiclib==1.52
geopy==2.2.0
idna==3.4
isort==5.10.1
mccabe==0.7.0
multidict==6.0.2
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
pip==22.2.2
pycodestyle==2.9.1
pydocstyle==6.1.1
pyflakes==2.5.0
python-box==6.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
setuptools==58.0.4
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.3
urllib3==1.26.12
voluptuous=0.13.1
xlrd=2.0.1
xlsxwriter-3.0.3
yarl=1.8.1